+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Hero component.
+- Hero1 component.
 - Hero2 component.
 - Footer component.
 - Plume Icon.